Update PredictResponse and TrainResponse DTOs to include HiddenSize property and adjust NumberCourses type

+++ StudentMarksPredictor.API/DTOs/PredictResponse.cs
@@ -2,7 +2,7 @@ namespace StudentMarksPredictor.API.DTOs;
 
 public class PredictResponse
 {
-    public double NumberCourses { get; set; }
+    public int NumberCourses { get; set; }
     public double TimeStudy { get; set; }
     public double PredictedMarks { get; set; }
     public Guid SessionId { get; set; }
